% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{BB}
\alias{BB}
\title{Bayesian bootstrap}
\usage{
BB(data, FUN, ..., B = 1000)
}
\arguments{
\item{data}{the original sample, as a vector, matrix or dataframe}

\item{FUN}{a function to calculate an estimate of the parameter of interest}

\item{B}{the number of bootstrap replications}
}
\value{
a list consisting of
\item{replicates}{a matrix containing an estimate for the parameter of interest of each bootstrap replicate}
\item{se}{the standard error(s) of the parameter estimates}
}
\description{
Bayesian bootstrap
}
\examples{
X <- rnorm(100, 4, 2)
BB_mean <- BB(X, weighted.mean, B=10000)
BB_mean$se
}

