% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{BLB.adapt}
\alias{BLB.adapt}
\title{BLB with adaptive selection on r and s}
\usage{
BLB.adapt(data, gamma, w_s, w_r, lambda, err, alpha)
}
\arguments{
\item{data}{a matrix or dataframe}

\item{gamma}{specifies the subsample size by b=n^gamma}

\item{w_s}{window size for the adaptive selection of s}

\item{w_r}{window size for the adaptive selection of r}

\item{lambda}{specifies the L2 penalty in Ridge regression}

\item{err}{relative error for the convergence criterion in adaptive selection}

\item{alpha}{the level of the confidence which is set to 0.05}
}
\value{
a list consisting of
\item{s}{the total number of subsamples from the original dataset}
\item{r}{the number of resamples for each subsample}
\item{mean_width}{mean_width the average marginal width of the confidence intervals across dimensions}
\item{mean_se}{mean_se the average standard error across the parameter dimensions}
}
\description{
BLB function which returns the average (across dimensions) confidence interal width or the average standard error of the parameter and the values of
r and s which were chosen adaptively within the function
}
\examples{
X <- mvrnorm(100 ,mu = rep(0,2) ,Sigma <- diag(2) )
epsilon <- rnorm(100,mean=0,sd=sqrt(10))
t_theta <- as.matrix( rep(1 ,2 ) )
Y <- X\%*\%t_theta + epsilon
my_data <- cbind(X ,Y )
my_result <- BLB_adapt(my_data ,0.7 ,3 ,20 ,10^(-5) ,0.05 ,alpha=0.05 )
}

