% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{BLB.multi}
\alias{BLB.multi}
\title{BLB for multivariate data}
\usage{
BLB.multi(data, gamma = 0.7, s = 20, r = 100, lambda = 10^-5,
  alpha = 0.05)
}
\arguments{
\item{data}{a matrix or dataframe}

\item{gamma}{specifies the size of the subsample by b=n^gamma}

\item{s}{number of subsamples}

\item{r}{number of resamples per subsample}

\item{lambda}{is the penalty in L2 norm for the ridge regression}

\item{alpha}{the level of the confidence for CI, respecified to 0.05}
}
\value{
a list consisting of
\item{output_CI}{average confidence intervals for the parameter estimate components}
\item{output_se}{average standard errors for the parameter estimates components}
}
\description{
BLB function that computes the (1-alpha)\% CI and the standard errors of the parameter estimate components
}
\examples{
n <- 100
d <- 5
X <- mvrnorm(n ,mu = rep(0 ,d ) ,Sigma <- diag( d ) )
epsilon <- rnorm(n ,mean = 0 ,sd = sqrt(10 ) )
t_theta <- as.matrix( rep(1 ,d ) )
Y <- X\%*\%t_theta + epsilon
data <- cbind(X ,Y )
my_result <- BLB.multi(data, 0,7, 15 ,100 ,10^(-5) )
}

