% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{BLB_adapt}
\alias{BLB_adapt}
\title{BLB with adaptive tuning parameters}
\usage{
BLB_adapt(data, gam, w_s, w_r, lam, err, alpha = 0.05)
}
\arguments{
\item{data}{a matrix or dataframe}

\item{gam}{specifies the subsample size b=n^gam}

\item{w_s}{window size for the adaptive selection of s}

\item{w_r}{window size for the adaptive selection of r}

\item{lam}{specifies the penalty in Ridge regression}

\item{err}{relative error for the convergence criterion in adaptive selection}

\item{alpha}{controls the width of the (1-alpha)\% confidence intervals}
}
\value{
a list consisting of
 \item{s}{the total number of subsamples from the original dataset}
 \item{r}{the number of resamples for each subsample}
 \item{mean_widths}{the average marginal width of the confidence intervals across dimensions}
}
\description{
BLB function which returns the average (across dimensions) confidence interal width of the parameter and the values of
r and s which were chosen adaptively
}
\examples{
X <- mvrnorm(100 ,mu = rep(0,2) ,Sigma <- diag(2) )
epsilon <- rnorm(100,mean=0,sd=sqrt(10))
t_theta <- as.matrix( rep(1 ,2 ) )
Y <- X\%*\%t_theta + epsilon
my_data <- cbind(X ,Y )
my_result <- BLB_adapt(my_data ,0.7 ,3 ,20 ,10^(-5) ,0.05 )
}

