% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap}
\usage{
bootstrap(data, FUN, ..., B = 1000)
}
\arguments{
\item{data}{the original sample, as a vector, matrix or dataframe}

\item{FUN}{a function to calculate an estimate of the parameter of interest. FUN must act on a vector in 1-dim case, or a matrix or data-frame in N-dim case. It must also return a single number.}

\item{B}{the number of bootstrap replications}

\item{data}{is a vector, matrix or data-frame}
}
\value{
a list consisting of
\item{replicates}{a matrix containing an estimate for the parameter of interest of each bootstrap replicate}
\item{se}{the standard error of the parameter estimates}
}
\description{
Bootstrap
}

