% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{cor_df}
\alias{cor_df}
\title{Pearson's correlation coefficient of bivariate data}
\usage{
cor_df(bivar_data)
}
\arguments{
\item{bivar_data}{A data-frame or matrix, the first two columns of which contain the two vectors for which the correlation coefficient is required}
}
\value{
the Pearson correlation coefficient of bivar_data[,1] and bivar_data[,2]
}
\description{
Pearson's correlation coefficient of bivariate data
}
\examples{
A <- data.frame(x=rnorm(10), y=rnorm(10, mean=2))
cor_df(A)
}

