% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Module1_code.R
\name{weighted.cor}
\alias{weighted.cor}
\title{Weighted correlation}
\usage{
weighted.cor(data, g)
}
\arguments{
\item{data}{a matrix or dataframe in which the first two columns contain the two vectors for which the weighted correlation is sought}

\item{g}{a vector of weights}
}
\value{
the weighted correlation
}
\description{
takes bivarate data and weights (g) and returns a weighted correlation
}
\examples{
df <- data.frame(norm = rnorm(10), unif = runif(10))
weights <- rep(1/10, 10)
weighted.cor(df, weights)
}

